In this branch I will attempt to merge the improvements made to SynEdit by the Letterpress project including eventually the Code folding stuff.
Letterpress is a major departure from the existing code base:
a) has removed SYN_CLX IFDEFs and support for earlier version of DELPHI using generics in a couple of places.
b) used the SynUnihighlighter as the basis for syntax highlighting and has dropped support for existing highlighters

The plan is to merge features one by one starting with the improved WordWrap plugin and printing later moving to Code Folding.